Rename "Refactoring" to "Maintenance"
"Refactoring" is a term of art in the Haskell community that has a subtly
different meaning outside the community.  Many people have more of an IDE
connotation when they hear refactoring, so I renamed this to "Maintenance" to
more accurately connote that Haskell is good at modifying and maintaining large
projects.